Move MainWindow and DialogBox to ui
