+++ src/main/java/luna/Main.java
@@ -7,6 +7,7 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
+import luna.ui.MainWindow;
 
 /**
  * A GUI for Luna using FXML.
