+++ src/main/java/luna/ui/DialogBox.java
@@ -1,4 +1,4 @@
-package luna;
+package luna.ui;
 
 import java.io.IOException;
 import java.util.Collections;
